Create K-Nearest_Neighbors (KNN)_algorithm _mplementation .c
The K-Nearest Neighbors (KNN) algorithm is a supervised machine learning technique used for classification and regression. It works by finding the K closest data points (neighbors) to a given input and uses majority voting (in classification) to determine the predicted label or class.

In this case, KNN is used to predict the type of jeans a person is likely to wear based on their age and monthly income in Nepali Rupees. Each data point represents a known person with labeled jeans type.